Large Update
 - Updated all files to work with the refactoring.
 - Added Extensions class.
 - AlexaComp.cs
    - Deleted some unused methods.
 - AlexaCompCore.cs
    - Moved splitStringEveryN to Core/Extensions.cs
    - Added region tags.
 - AlexaCompHARDWARE.cs
    - Added region tags.
    - Deleted SensorObject
 - AlexaCompREQUEST.cs
    - Moved Request and Response classes to AlexaComp/Core
 - AlexaCompSERVER.cs
    - Added region tags.
 - LightingController.cs
    - Removed Extensions class.
