+++ AlexaComp/AlexaComp.cs
@@ -11,6 +11,8 @@ using System.Text.RegularExpressions;
 using System.Xml.XPath;
 using AlexaComp.Controllers;
 using AlexaComp.Forms;
+using AlexaComp.Core;
+using System.Management;
 
 using log4net;
 using log4net.Config;
@@ -47,6 +49,7 @@ namespace AlexaComp {
                     }
                 }
             }
+
             // Log Paths
             _log.Info(exePath.ToString());
             _log.Info("PathToDebug - " + pathToDebug);
@@ -60,8 +63,10 @@ namespace AlexaComp {
             clog("Exception Handler Registered");
 
             getExternalIP();
+            Console.ReadLine();
+            // testThing();
 
-            LoadingScreenThread.Start();
+            // LoadingScreenThread.Start();
             /*
             LightingControlThread.Start();
             Console.WriteLine("Waiting Before Effect Start");
@@ -94,78 +99,6 @@ namespace AlexaComp {
             } else {
                 clog("IP Regex Match Unsuccessful.", "ERROR");
             }
-        }
-        
-        /*
-        * Verifies that the necessary config values are entered and correctly formatted.
-        */
-        public static void verifyConfig() {
-            foreach (KeyValuePair<string, string> pair in settingsDict){
-                // Console.WriteLine("pair: '" + pair.Key + "' - " + pair.Value);
-                if (String.IsNullOrEmpty(pair.Value)){
-                    _log.Fatal("Config Value" + pair.Key + " not configured correctly or are missing. AlexaComp will not work.");
-                } else if (pair.Key == "HOST"){ // Is Host IP formatted properly?
-                    bool isValidIp = IPAddress.TryParse(pair.Value, out IPAddress ipAdd);
-                    if (isValidIp != true){
-                        _log.Fatal("Config value \"HOST\" exists, but is not configured or formatted properly.");
-                    }
-                } 
-            }
-        }
-        
-        /*
-        * Updates a config value of name {key} with value {value}.
-        * @ param key - The name of the config value to update.
-        * @ param value - The value to update to.
-        */
-        public static void UpdateConfigValue(string key, string value){
-            Console.WriteLine(key + " - " + value);
-            ExeConfigurationFileMap debugMap = new ExeConfigurationFileMap { ExeConfigFilename = pathToDebug + "\\AlexaComp.exe.config" };
-            ExeConfigurationFileMap projectMap = new ExeConfigurationFileMap { ExeConfigFilename = pathToProject + "\\App.config" };
-            var config = ConfigurationManager.OpenMappedExeConfiguration(debugMap, ConfigurationUserLevel.None);
-            var config2 = ConfigurationManager.OpenMappedExeConfiguration(projectMap, ConfigurationUserLevel.None);
-            config.AppSettings.Settings.Remove(key);
-            config2.AppSettings.Settings.Remove(key);
-            config.AppSettings.Settings.Add(key, value);
-            config2.AppSettings.Settings.Add(key, value);
-            config.Save(ConfigurationSaveMode.Modified);
-            config2.Save(ConfigurationSaveMode.Modified);
-
-            ConfigurationManager.RefreshSection("appSettings");
-        }
-
-        /*
-        * Inventories all user installed programs.
-        */
-        public static void inventoryPrograms() {
-            string registry_key = @"SOFTWARE\Microsoft\Windows\CurrentVersion\Uninstall";
-            using (Microsoft.Win32.RegistryKey key = Registry.LocalMachine.OpenSubKey(registry_key)) {
-                foreach (string subkey_name in key.GetSubKeyNames()) {
-                    using (RegistryKey subkey = key.OpenSubKey(subkey_name)) {
-                        if (subkey.GetValue("DisplayName") != null) {
-                            string subkeyValue = subkey.GetValue("DisplayName").ToString();
-                            if (subkeyValue.Contains("Microsoft")  || subkeyValue.Contains("SDK") || 
-                                subkeyValue.Contains("WinRT") || subkeyValue.Contains("vs_") ||
-                                subkeyValue.Contains("Visual C++") || subkeyValue.Contains("CRT") ||
-                                subkeyValue.Contains("Windows")){
-                                
-                            } else {
-                                Console.WriteLine(subkey.GetValue("DisplayName"));
-                            }
-                        }
-                    }
-                }
-            }
-        }
-    }
-
-    class Options {
-        public Request req;
-        public System.Net.Sockets.NetworkStream nwStream;
-
-        public Options(Request req_, System.Net.Sockets.NetworkStream nwStream_) {
-            req = req_;
-            nwStream = nwStream_;
         }
     }
 }
