+++ AlexaComp/AlexaComp.csproj
@@ -169,6 +169,9 @@
   </ItemGroup>
   <ItemGroup>
     <Compile Include="AlexaCompCore.cs" />
+    <Compile Include="Core\Extensions.cs" />
+    <Compile Include="Core\Request.cs" />
+    <Compile Include="Core\Response.cs" />
     <Compile Include="Forms\AdvancedSettingsForm.cs">
       <SubType>Form</SubType>
     </Compile>
@@ -185,8 +188,8 @@
     <Compile Include="AlexaCompSERVER.cs" />
     <Compile Include="AlexaCompREQUEST.cs" />
     <Compile Include="AlexaCompHARDWARE.cs" />
-    <Compile Include="Controllers\AudioController.cs" />
-    <Compile Include="Controllers\LightingController.cs" />
+    <Compile Include="Core\Controllers\AudioController.cs" />
+    <Compile Include="Core\Controllers\LightingController.cs" />
     <Compile Include="Forms\LoadingScreenForm.cs">
       <SubType>Form</SubType>
     </Compile>
@@ -195,6 +198,7 @@
     </Compile>
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="Tests\TestFunctions.cs" />
+    <Compile Include="Core\ProgramInventory.cs" />
     <EmbeddedResource Include="Forms\AdvancedSettingsForm.resx">
       <DependentUpon>AdvancedSettingsForm.cs</DependentUpon>
     </EmbeddedResource>
@@ -263,6 +267,9 @@
       <IncludeHash>True</IncludeHash>
       <FileType>File</FileType>
     </PublishFile>
+  </ItemGroup>
+  <ItemGroup>
+    <Folder Include="SubProjects\" />
   </ItemGroup>
   <Import Project="$(MSBuildToolsPath)\Microsoft.CSharp.targets" />
   <Import Project="packages\RGB.NET.Resources.Asus.0.1.0\build\RGB.NET.Resources.Asus.targets" Condition="Exists('packages\RGB.NET.Resources.Asus.0.1.0\build\RGB.NET.Resources.Asus.targets')" />
