+++ AlexaComp/AlexaComp.sln
@@ -5,10 +5,6 @@ VisualStudioVersion = 15.0.28010.2036
 MinimumVisualStudioVersion = 10.0.40219.1
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "AlexaComp", "AlexaComp.csproj", "{B8379353-325F-4858-9E4A-0E9FD4357872}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "CompatibilityTool", "CompatibilityTool\CompatibilityTool.csproj", "{85382666-83F7-48BC-A3C7-9180B7A8FDF9}"
-EndProject
-Project("{54435603-DBB4-11D2-8724-00A0C9A8B90C}") = "AlexaCompSetup", "AlexaCompSetup\AlexaCompSetup.vdproj", "{280F63E6-6C3B-4B3F-90FD-31E8AD831A3C}"
-EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -19,12 +15,6 @@ Global
 		{B8379353-325F-4858-9E4A-0E9FD4357872}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{B8379353-325F-4858-9E4A-0E9FD4357872}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{B8379353-325F-4858-9E4A-0E9FD4357872}.Release|Any CPU.Build.0 = Release|Any CPU
-		{85382666-83F7-48BC-A3C7-9180B7A8FDF9}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{85382666-83F7-48BC-A3C7-9180B7A8FDF9}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{85382666-83F7-48BC-A3C7-9180B7A8FDF9}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{85382666-83F7-48BC-A3C7-9180B7A8FDF9}.Release|Any CPU.Build.0 = Release|Any CPU
-		{280F63E6-6C3B-4B3F-90FD-31E8AD831A3C}.Debug|Any CPU.ActiveCfg = Debug
-		{280F63E6-6C3B-4B3F-90FD-31E8AD831A3C}.Release|Any CPU.ActiveCfg = Release
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
