+++ AlexaComp/Forms/AdvancedSettingsForm.cs
@@ -7,7 +7,9 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 using System.Windows.Forms;
+
 using AlexaComp;
+using AlexaComp.Core;
 
 namespace AlexaComp.Forms {
     public partial class AdvancedSettingsForm : Form {
